server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Charith <114335319+Charith0401@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"  
asgiref = "3.8.1"
django = "5.0.6"
django-cors-headers = "4.3.1"
djangorestframework = "3.15.1"
psycopg2 = "2.9.9"
sqlparse = "0.5.0"
tzdata = "2024.1"
drf-spectacular="0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"